sync support

+++ entrypoint.sh
@@ -14,8 +14,13 @@ echo "[default]
 retry_mode = standard
 max_attempts = 3" > ~/.aws/config
 
+if [ -f "${FILE}" ] ; then
 aws s3 cp ${FILE} s3://${AWS_S3_BUCKET}${AWS_S3_PATH} \
-            --region ${AWS_REGION} $*
+            --region ${AWS_REGION} $*    
+else
+aws s3 sync ${FILE} s3://${AWS_S3_BUCKET}${AWS_S3_PATH} \
+            --region ${AWS_REGION} $*    
+fi
 
 status=$?
